#2. Separate Program into WSA Blocking Server and Client

+++ GSP_Client/GSP_Client/Board.cpp
@@ -4,7 +4,7 @@ Board::Board() {
     for (int i = 0; i < 8; ++i) {
         for (int j = 0; j < 8; ++j) {
             squares_[i * 8 + j].setSize(sf::Vector2f(SQUARE_SIZE, SQUARE_SIZE));
-            squares_[i * 8 + j].setPosition(j * SQUARE_SIZE, i * SQUARE_SIZE);
+            squares_[i * 8 + j].setPosition(static_cast<float>(j * SQUARE_SIZE), static_cast<float>(i * SQUARE_SIZE));
             if ((i + j) & 1) squares_[i * 8 + j].setFillColor(sf::Color::White);
             else squares_[i * 8 + j].setFillColor(sf::Color::Black);
         }
